Chapter 4: fix east-column displacement — push() still truncated wx,wy to int16
The 8x8 world reaches 40960 subcells, past int16's 32767. render.h's Inst.wx,wy
were widened to int32 (and the vertex layout + WGSL updated), but push() still cast
the value to (int16_t) BEFORE storing it into the int32 field. So every placement
east of x=32767 wrapped to a large negative and drew far to the west — leaving only
~6.5 of the 8 screen-columns in place, the rest displaced as a floating strip on the
opposite side (most visible at low zoom + steep pitch). The earlier int32 migration
was incomplete: it widened the field but missed the cast.

Store wx,wy without the 16-bit cast. The other fields stay int16 (small heights /
half-extents in subcells, facing in Q14).

Add a regression test: in the whole-world emit, every wx,wy must stay within
[-SUB, ARENA+SUB] (a margin for edge barrels that poke just past a wall) — i.e. no
int16 wrap. It fails on the old cast (a wx hit ~-32752) and passes now. 147 checks.

+++ tank-combat/chapter-4/src/render.c
@@ -106,7 +106,11 @@ static uint32_t push(Inst* out, uint32_t k, int cx, int cy, int cz,
                      int hx, int hy, int hz, int co, int si, uint32_t rgba) {
   if (k >= INST_MAX) return k;
   Inst* o = &out[k];
-  o->wx = (int16_t)cx; o->wy = (int16_t)cy; o->wz = (int16_t)cz; o->hz = (int16_t)hz;
+  /* wx,wy are int32 — the 8x8 arena reaches 40960 subcells, past int16. Store them
+   * WITHOUT a 16-bit cast (the old (int16_t) truncation wrapped the east columns to
+   * large negatives, drawing them as a displaced strip far to the west). The remaining
+   * fields are int16 and small (heights/half-extents in subcells, facing in Q14). */
+  o->wx = (int32_t)cx; o->wy = (int32_t)cy; o->wz = (int16_t)cz; o->hz = (int16_t)hz;
   o->hx = (int16_t)hx; o->hy = (int16_t)hy; o->co = (int16_t)co; o->si = (int16_t)si; o->rgba = rgba;
   return k + 1;
 }
